+++ cookbooks/devel/recipes/bower.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: devel
 # Recipe:: bower
 #
-# Copyright 2013,2014 Carles Muiños
+# Copyright 2013-2015 Carles Muiños
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,10 +18,8 @@
 # limitations under the License.
 #
 
-
 # A package manager for the web
-execute "bower-installation" do
-  command "npm install -g bower"
-  not_if { ::File.exists? "/usr/bin/bower" }
+execute 'bower-installation' do
+  command 'npm install -g bower'
+  not_if { ::File.exist? '/usr/bin/bower' }
 end
-