+++ cookbooks/devel/recipes/ruby.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: devel
 # Recipe:: ruby
 #
-# Copyright 2013,2014 Carles Muiños
+# Copyright 2013-2015 Carles Muiños
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,7 +17,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
 
 # refs:
 #   https://robots.thoughtbot.com/using-rbenv-to-manage-rubies-and-gems
@@ -36,18 +35,18 @@ ruby = devel['profiles']['ruby']
 
 if app_available? ruby
   # Installing curl ...
-  package "curl"
+  package 'curl'
 
   # Installing rbenv, to change Ruby versions ...
-  install_app "ruby" do
+  install_app 'ruby' do
     force true
     profile ruby
   end
 
   # Setting rbenv environment up ...
-  bootstrap "ruby" do
-    folders [".rbenv/plugins"]
-    env :priority => "10"
+  bootstrap 'ruby' do
+    folders ['.rbenv/plugins']
+    env priority: '10'
   end
 
   box = node[:box]
@@ -57,8 +56,8 @@ if app_available? ruby
 
     # Installing rbenv-binstubs to make rbenv transparently aware of project-specific binstubs created by bundler ...
     git "#{username}-rbenv-binstubs" do
-      repository "https://github.com/ianheggie/rbenv-binstubs.git"
-      revision "master"
+      repository 'https://github.com/ianheggie/rbenv-binstubs.git'
+      revision 'master'
       destination "#{usr[:home]}/.rbenv/plugins/rbenv-binstubs"
       user username
       group usr[:group]
@@ -67,8 +66,8 @@ if app_available? ruby
 
     # Installing rbenv-gem-rehash so the shell automatically picks up binaries after installing gems with binaries ...
     git "#{username}-rbenv-gem-rehash" do
-      repository "https://github.com/sstephenson/rbenv-gem-rehash.git"
-      revision "master"
+      repository 'https://github.com/sstephenson/rbenv-gem-rehash.git'
+      revision 'master'
       destination "#{usr[:home]}/.rbenv/plugins/rbenv-gem-rehash"
       user username
       group usr[:group]
@@ -77,8 +76,8 @@ if app_available? ruby
 
     # Installing rbenv-env, adds the rbenv env command that shows relevant environment variables ...
     git "#{username}-rbenv-env" do
-      repository "https://github.com/ianheggie/rbenv-env.git"
-      revision "master"
+      repository 'https://github.com/ianheggie/rbenv-env.git'
+      revision 'master'
       destination "#{usr[:home]}/.rbenv/plugins/rbenv-env"
       user username
       group usr[:group]
@@ -87,8 +86,8 @@ if app_available? ruby
 
     # Installing rbenv-update, adds the rbenv update command that updated rbenv and all installed plugins ...
     git "#{username}-rbenv-update" do
-      repository "https://github.com/rkh/rbenv-update.git"
-      revision "master"
+      repository 'https://github.com/rkh/rbenv-update.git'
+      revision 'master'
       destination "#{usr[:home]}/.rbenv/plugins/rbenv-update"
       user username
       group usr[:group]
@@ -97,8 +96,8 @@ if app_available? ruby
 
     # Installing ruby-build, to install Rubies ...
     git "#{username}-ruby-build" do
-      repository "https://github.com/sstephenson/ruby-build.git"
-      revision "master"
+      repository 'https://github.com/sstephenson/ruby-build.git'
+      revision 'master'
       destination "#{usr[:home]}/.rbenv/plugins/ruby-build"
       user username
       group usr[:group]
@@ -141,4 +140,3 @@ if app_available? ruby
     end
   end
 end
-