+++ cookbooks/devel/recipes/shelr.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: devel
 # Recipe:: shelr
 #
-# Copyright 2013,2014 Carles Muiños
+# Copyright 2013-2015 Carles Muiños
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,11 +18,9 @@
 # limitations under the License.
 #
 
-
 devel = node[:apps][:devel]
 
 # Utility for plain text screencasting
-install_app "shelr" do
+install_app 'shelr' do
   profile devel['profiles']['shelr']
 end
-