+++ cookbooks/devel/recipes/wsk.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: devel
 # Recipe:: wsk
 #
-# Copyright 2013,2014 Carles Muiños
+# Copyright 2013-2015 Carles Muiños
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,14 +18,13 @@
 # limitations under the License.
 #
 
-
 devel = node[:apps][:devel]
 
 # Boilerplate & tooling for multi-device development
 wsk = devel['profiles']['wsk']
 
 if app_available? wsk
-  install_app "wsk" do
+  install_app 'wsk' do
     force true
     profile wsk
   end
@@ -61,4 +60,3 @@ if app_available? wsk
     end
   end
 end
-